First draft. butler font, sass, index.html, app.js and favicon
It is now basically text with some weak styling. I don't want this
too cluttered, so there will not be too much styling.
There is a copy-button that copies the email text to clipboard.
The editable field has a red underline. Need to explain it.
The email only has a blue border. It needs something more.
The text need to be changed. And we need swedish.